Add a slice-like object that can be used to edit in place
While copying is pretty fast, if we can do something without copying
that would naturally be faster. Thankfully we can use Rust's ownership
mechanics and Index traits to build a slice-like object that can let a
user edit the data and then copy the result without any unintended side
effects. Since the current flow is to copy, edit, and copy into the
next buffer, we can do that first "copy" in place in the memory of the
input buffer. Once this returned slice leaves scope the data will no
longer be accessible and doesn't interfere with yet read data.