Control packet size and add tcp connections and the chrome socket
Pause the child.stdout stream and manually read from it to reduce the
default 8192 byte buffer to whatever we want (192). Changing the
buffering on this side from 42ms to 1ms (channels * byte width * time).

Tcp socket on rust side exposes an api to control stream blocking
behaviour, this may a performance improvement there. Or the hyper
interface may be obfuscating the fact that it just returns Ok(0) when
it would otherwise block.