Replace graph_utils::RingBuffer's guts with iterators
This removes some duplicated code and improves performance a lot as it
removes or moves a lot of repeated logic to instead happen once or
never. For example in the case of writing to buffer, the end_index is
moved once instead of for every item in the incoming buffer.